% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/align.3D.img.in.2D.R
\name{align.3D.img.in.2D}
\alias{align.3D.img.in.2D}
\title{Display a 3d image on screen in 3D by tiling it.}
\usage{
align.3D.img.in.2D(img = NULL, dimension = 1)
}
\arguments{
\item{img}{the image to put into 3D}

\item{dimension}{the dimension to reduce across}
}
\description{
Display a 3d image on screen in 3D by tiling it.
}

