% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/caret.train.model.list.formula.R
\name{caret.train.model.list.formula}
\alias{caret.train.model.list.formula}
\title{Can take x and y values as the defaults.}
\usage{
\method{caret.train.model.list}{formula}(formula, data, trControl,
  training.list, ...)
}
\arguments{
\item{formula}{}

\item{data}{}

\item{trControl}{passed directly to train {caret}; a list of values that define how this function acts. Default value if each item doesn't have its own trControl. See trainControl and http://topepo.github.io/caret/training.html#custom. (NOTE: If given, this argument must be named.)}

\item{training.list;}{a list describing a list of train {caret} values to run. Should be a list of objects generated by get.caret.model.spec. Each should contain exactly two values, to be passed to train {caret}: method and tuning. If tuning is an integer, it will be passed to tuneLength. If tuning is a data frame, it will be passed to tuneGrid. If it is null, train's default values for tuneLength will apply. Otherwise an error is generated.}
}
\description{
Can take x and y values as the defaults.
}

