% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/createFoldsByGroup.R
\name{createFoldsByGroup}
\alias{createFoldsByGroup}
\title{Intended for use with caret; creates folds based on the group allocations
Use in place of createFolds}
\usage{
createFoldsByGroup(group.allocation)
}
\arguments{
\item{group.allocation}{a vector describing which group/fold each item belongs to.}
}
\description{
Intended for use with caret; creates folds based on the group allocations
Use in place of createFolds
}
\examples{
createFoldsByGroup(c(1,1,1,1,2,2,2,2,3,3,3))
#create 3 folds; two with 4 members each and the third with 3 members.
}

