% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/display.2D.img.R
\name{display.2D.img}
\alias{display.2D.img}
\title{Display a 2d image on screen.}
\usage{
display.2D.img(ds)
}
\description{
Display a 2d image on screen.
}
\examples{
display.2D.img(array(rnorm(100^2),dim=c(100, 100)))
}

