% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/display.3D.img.in.2D.R
\name{display.3D.img.in.2D}
\alias{display.3D.img.in.2D}
\title{Display a 3d image on screen in 3D by tiling it.}
\usage{
display.3D.img.in.2D(img, dimension = 3)
}
\description{
Display a 3d image on screen in 3D by tiling it.
}
\examples{
data <- array(sample(1:100,10^3,replace=TRUE),c(10,10,10))
display.3D.img.in.2D(data,1)
}

