% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.roi.R
\name{extract.roi}
\alias{extract.roi}
\title{Extract an ROI from an image timeseries using a mask. Expects an image in the format of NIFTI file. If it's not, we get trouble.}
\usage{
extract.roi(img.ts, roi.mask)
}
\arguments{
\item{img.ts}{a 4D image timeseries from which to extract the mask}

\item{roi.mask}{the roi mask to use to extract the file.}
}
\description{
Extract an ROI from an image timeseries using a mask. Expects an image in the format of NIFTI file. If it's not, we get trouble.
}
\examples{
image.filename <- system.file("extdata", "haxby2001subj1bold.nii.gz", package = "mvpa")
mask.filename <- system.file("extdata", "haxby2001subj1mask.nii.gz", package = "mvpa")
fmri.image <- read.image(image.filename)
mask.image <- read.image(mask.filename)
roi.data <- extract.roi(fmri.image,mask.image)
summary(roi.data)
}

