% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.caret.model.spec.R
\name{get.caret.model.spec}
\alias{get.caret.model.spec}
\title{Intended for use with caret; specifies a caret model specification for use by companion function
caret.train.model.list
There is a list of models here: http://topepo.github.io/caret/modelList.html}
\usage{
get.caret.model.spec(method, tuning = NULL, preProcess = NULL, ...)
}
\arguments{
\item{method}{method to be passed to train {caret}}

\item{tuning}{tuning method to be passed to train {caret} etiher a tuneGrid data frame or an integer to be passed to tuneLength}

\item{...}{Values that will be passed directly to the function; in caret, these are values that aren't supported as tuning parameters.}
}
\description{
Intended for use with caret; specifies a caret model specification for use by companion function
caret.train.model.list
There is a list of models here: http://topepo.github.io/caret/modelList.html
}
\examples{
get.caret.model.spec("knn",tuning = 5, preProcess = "pca")
#get a knn model spec, with 5 a tuneLength of 5, and use PCA pre-processing.
}

