% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load.preprocess.subject.R
\name{load.preprocess.subject}
\alias{load.preprocess.subject}
\title{load and pre-process a subject}
\usage{
load.preprocess.subject(img.path, mask.path, run.path = NULL,
  run.factor.list = NULL)
}
\arguments{
\item{run.path}{list of the run objects as a file to look up}

\item{run.factor.list}{should be a data frame, the first column of which is called "labels"
and describes the labels applied to each image; the second column called "chunks" and describes any applicable chunks (e.g., runs)}

\item{x}{fmri timeseries}
}
\description{
load and pre-process a subject
}

