\name{mvpa-package}
\alias{mvpa-package}
\alias{mvpa}
\docType{package}
\title{
Multi-voxel pattern analysis
}
\description{
Provides shortcuts to streamline MVPA using the R caret package
}
\details{
\tabular{ll}{
Package: \tab mvpa\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-29\cr
License: \tab GPL 3.0\cr
}
This package streamlines testing MVPA solutions in R using caret.
caret.train.model.list takes a list of different caret functions and trains the data on each of them, passing the result back in a single list so results can be easily compared.
}
\author{
This package was authored by Ben Smith.

Maintainer: Ben Smith <benjsmith@gmail.com>
}
\references{
}
\keyword{ mvpa }
\seealso{
}
\examples{
}
