% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepropress.fmri.run.set.R
\name{prepropress.fmri.run.set}
\alias{prepropress.fmri.run.set}
\title{Apply preprocessing stuff.}
\usage{
prepropress.fmri.run.set(x, run)
}
\arguments{
\item{x}{fmri timeseries}

\item{run}{list of the run objects}
}
\description{
Apply preprocessing stuff.
}

