% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepropress.haxby.run.set.R
\name{prepropress.haxby.run.set}
\alias{prepropress.haxby.run.set}
\title{Apply preprocessing stuff.}
\usage{
prepropress.haxby.run.set(x, run)
}
\arguments{
\item{file}{}

\item{file.format}{file format the mri file is stored in. Currently only NIFTI}
}
\description{
Apply preprocessing stuff.
}

