% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.image.R
\name{read.image}
\alias{read.image}
\title{Read a brain image
http://hilaryparker.com/2014/04/29/writing-an-r-package-from-scratch/
This function allows you to read a brain image very nicely}
\usage{
read.image(file, dim, file.format = "NIFTI")
}
\arguments{
\item{file}{}

\item{file.format}{file format the mri file is stored in. Currently only NIFTI}
}
\description{
Read a brain image
http://hilaryparker.com/2014/04/29/writing-an-r-package-from-scratch/
This function allows you to read a brain image very nicely
}

