% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/register.machine.cores.R
\name{register.machine.cores}
\alias{register.machine.cores}
\title{Detects and registers the number of machine cores. This should be run before using cross-validation with caret
Will allow R to use the doMC package to utilize the machine's multiple processors.
You will need to call this function for it to be applied.}
\usage{
register.machine.cores()
}
\description{
Detects and registers the number of machine cores. This should be run before using cross-validation with caret
Will allow R to use the doMC package to utilize the machine's multiple processors.
You will need to call this function for it to be applied.
}
\examples{
register.machine.cores()
}

