% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/read.image.R
\name{read.image}
\alias{read.image}
\title{Read a brain image}
\usage{
read.image(file, file.format = "NIFTI")
}
\arguments{
\item{file}{}

\item{file.format}{file format the mri file is stored in. Currently only NIFTI}
}
\description{
This function allows you to read a brain image
}
\examples{
read.image()
}

